Standalone PKGBUILD
Changed my mind about a separate install script. It'd be confusing if users end
up with multiple me3 installs in ~ and /usr/local before they even get to a
distro package. Also it should have an uninstaller; at that point I'm
reinventing make or just writing a Makefile.

Another advantage is this PKGBUILD can be built immediately for me3 0.7.0 and
doesn't need a new release.